Modified: tsp.h, utilities.c, tsp.c
In tsp.c:
all the function prototypes have been mooved in tsp.h, so when you include the file all the functions defined there can be used.

In utilities.c:
the print_plot function has been updated, now it writes in the file to plot also the non-zero optimal variables.

In tsp.c:
the TSPopt function has been updated, now it computes (and print) the best solution according to the model built, then it saves the values of the all variables in inst->best_sol.